code cleanup
code has been cleaned up a bit. the test index has been split up and all parser related tests have their own script now.
